refactor:자동스크롤링 훅으로 분리
